+++ NEWS.md
@@ -1,5 +1,6 @@
 # santoku (development version)
 
+* You can now set `labels = NULL` to return integer codes.
 * Bugfix: `lbl_endpoint()` was erroring for some types of breaks.
 * Use unicode em-dash and no spaces in `lbl_dash()` and `lbl_discrete()`.
 * The `groups` argument to `chop_evenly()`,  deprecated in 0.4.0, has